Separate components to make it more testable
