+++ packages/react-components/lib/map/three-fiber/index.tsx
@@ -6,3 +6,4 @@ export * from './robot-three-maker';
 export * from './shape-three-rendering';
 export * from './image-maker';
 export * from './text-maker';
+export * from './circle-shape';
